requirements.txt
pytest
mock
coverage
tk
Pillow
flake8
